use compound instead of inheritance
The inheritance scheme was proofing to give a lot subtle errors
when using and compiling different test scenario's. Which proofs
the best practice of not inheriting from std containers.

Moving to a compound pattern, seems to be the better choice if
we don't want to use free-functions but dedicated containers,
for Points and Polygons.

This does require a lot of boilerplate code to follow to
the practice of the least surprises.